Refactor inference streaming after code review follow-ups.
Wire within-tier incremental solution admission, extract accelerated inference into a shared module, unify halt terminal row completion, and move frontend stream state into a typed reducer with parser-safe arithmetic mapping.
