+++ packages/api/api/analytics/military_score_inference/inference_stream_domain_events.py
@@ -53,9 +53,5 @@ class GlobalPauseChanged:
 
 
 InferenceStreamDomainEvent = (
-    HeldSolutionsUpdated
-    | TierProgress
-    | RowComplete
-    | RowFailed
-    | GlobalPauseChanged
+    HeldSolutionsUpdated | TierProgress | RowComplete | RowFailed | GlobalPauseChanged
 )
