+++ packages/frontend/src/analytics/scores/inferenceRowStreamState.test.ts
@@ -1,0 +1,49 @@
+import { describe, expect, it } from 'vitest'
+import {
+  initialRowStreamState,
+  reduceRowStreamState,
+  rowDetailFromStreamState,
+  stablePlayerIdsKey,
+} from './inferenceRowStreamState'
+
+describe('stablePlayerIdsKey', () => {
+  it('sorts ids so order changes do not alter the key', () => {
+    expect(stablePlayerIdsKey([9, 8])).toBe('8,9')
+    expect(stablePlayerIdsKey([8, 9])).toBe('8,9')
+  })
+})
+
+describe('reduceRowStreamState', () => {
+  it('replaces held solutions wholesale on solution events', () => {
+    const state = initialRowStreamState()
+    const next = reduceRowStreamState(state, {
+      type: 'solution',
+      solutions: [
+        {
+          objectiveValue: 10,
+          actions: [{ actionId: 'a1', label: 'Build fighter', count: 1 }],
+        },
+      ],
+    })
+
+    expect(next.heldSolutions).toHaveLength(1)
+    expect(rowDetailFromStreamState(8, next).solutionCount).toBe(1)
+    expect(rowDetailFromStreamState(8, next).displayStatus).toBe('success')
+  })
+
+  it('marks paused rows without clearing held solutions', () => {
+    const withSolution = reduceRowStreamState(initialRowStreamState(), {
+      type: 'solution',
+      solutions: [
+        {
+          objectiveValue: 10,
+          actions: [{ actionId: 'a1', label: 'Build fighter', count: 1 }],
+        },
+      ],
+    })
+    const paused = reduceRowStreamState(withSolution, { type: 'globalPause', paused: true })
+
+    expect(paused.status).toBe('paused')
+    expect(paused.heldSolutions).toHaveLength(1)
+  })
+})
